alx_travel_app/alx_travel_app/alx_travel_app/requirement.txt
Django>=4.2,<5.0
djangorestframework
django-cors-headers
drf-yasg
django-environ
celery
mysqlclient
